Bind val to suggest callback to avoid race
When the 'suggest' function is passed to the 'source' callback,
there is a race condition that can result in caching the suggested
completions against a value that has since been updated.  Instead,
we add the current value for which we are asking for suggestions
as a parameter to the 'suggest' function and bind the value at the
time of making the 'source' callback.