Create an event to manually trigger the search
It may be useful to ask this plugin to show it's suggestions immediately, perhaps from a virtual keyboard that doesn't send real keydown events or some other trigger.  This change creates a new 'autocomplete.popup' event that can be triggered through jQuery's trigger() routine to request it immediately do the search and reveal the results.